Very simple tsv reader design csv_reader.
Only reads a line of a space-separated variable file, tokenizes the
line, then converts the string tokens to floats, then prints out the
line. Need to make all the lines stored rather than overwritten. This is
a prototype for a function in the main program for reading in coil and
magnetic data.