Added compiler warnings with -Wall again.
There's a warning about Boundary::CalcB not having a return statement
since it's a nonvoid function. I suppose in practice we will always want
to use SlowBoundary and not Boundary but for now I can't make
Boundary::CalcB a pure virtual since one is created in elliptic_test.
Should elliptic_test be updated?