Broke out two subroutines from GS_Solver.
The jphi initialization is now in a subroutine.
The Grad_Output choosing and initialization is now in a subroutine.

Some variables from the config file are only used once in some
initialization; I made it so instead of having another variable that is
only used once, it is taken from vm directly into the function.

I made the tabs two spaces instead of four, which is more
consistent.